==> notebooks/04-imputation.Rmd <==
---
title: "scRNA-Seq Anlaysis - 04 - Imputation"
output:
  github_document:
    toc: true
---

#Dependencies
```{r}
library(scater)
```

#Imputation
Here we'll use MAGIC (Van Dijk, BioRxiv, 2017) to smooth the data and restore some of the structure in the data that's lost to drop-out and other sources of technical variation. 

The most developed implementation of MAGIC right now is in python, so I'll move the data over there and then bring it back into R.

Note: After running it on exprs.norm in it's current state, cell cycle started to affect things. Given the small number of proliferating cells, I'll just eliminate S/G2M cells from downstream analysis. Also, IFC column 13 really stands out as an outlier (even in current exprs.norm), so I'll get rid of them too.

```{r}
sce <- readRDS("../output/sce.normalized.rds")
sce <- scater::filter(sce, CellCycle=="G1")
sce <- scater::filter(sce, IFC.Column!=13)
write.csv(t(assay(sce, "exprs.norm")), file="../output/exprs.norm.before.imputation.csv", quote=F, row.names=T)
```

Jump over to 04-imputation-magic.ipynb for the MAGIC script.

#Loading imputed data back in
```{r}
magic <- read.csv("../output/exprs.norm.after.imputation.csv", row.names=1)
colnames(magic) <- rownames(sce)
assay(sce, "exprs.magic") <- t(as.matrix(magic))
```

#Assessing the imputation
```{r, fig.height=3.4, fig.width=4.75}
plotPCA(sce, exprs_values="exprs.norm", shape_by="Condition", 
        colour_by="Condition")
plotPCA(sce, exprs_values="exprs.magic", shape_by="Condition", 
        colour_by="Condition")
```

Imputation really seems to bring out the divergence in the transcriptional manifolds between the two conditions.

```{r, fig.height=3.4, fig.width=4.75}
plotPCA(sce, exprs_values="exprs.magic", shape_by="Condition", 
        colour_by="Condition", ncomponents=3)
```

# Save point
```{r}
saveRDS(sce, file="../output/sce.normalized.imputed.rds")
```
